[refac/WEDPRO-190] refac: 상품 등록 모달 모달에 상태에따른 로직 분리
