+++ src/components/goodsmodal/GoodsModal.jsx
@@ -1,274 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
 import { AiOutlineClose } from "@react-icons/all-files/ai/AiOutlineClose";
-import {
-  postGoodsProductApi,
-  updateGoodsname,
-  updateGoodsPrice,
-} from "../../services/goods/GoodsProductService";
+import { postGoodsProductApi } from "../../services/goods/GoodsProductService";
 
+import CreateState from "./createState/CreateState";
 import logo from "@/assets/icons/logo.png";
-
-//등록 상태
-//상품 등록 전에 빈 text랑 등록 확인 취소
-// 등록후 값이 변경되고 등록 확인 취소 버튼만 보인다.
-function CreateGoodsState({
-  getGoodsUrlItem,
-  setGetGoodsUrlItem,
-  setIsOpen,
-  postGoodsListRender,
-  goodsData,
-  deleteGoodsRender,
-}) {
-  const getGoodsUrl = (e) => {
-    setGetGoodsUrlItem(e.target.value);
-  };
-
-  const okButton = () => {
-    setIsOpen(false);
-  };
-
-  const deleteButton = async (id) => {
-    await deleteGoodsRender(id);
-    setIsOpen(false);
-  };
-
-  const registerGoodsButton = async () => {
-    await postGoodsListRender(getGoodsUrlItem);
-  };
-  return (
-    <>
-      <GoodsDiv>
-        {goodsData?.length !== 0 && goodsData ? (
-          <div id={goodsData.usersGoodsId} key={goodsData.usersGoodsId}>
-            <CreateGoodsImage url={goodsData.usersGoodsImgUrl} />
-            <div>
-              <GoodsDonationDiv>
-                <p>상품 이름 :{goodsData.usersGoodsName}</p>
-                <GoodsText>
-                  후&nbsp; 원 &nbsp; 가 : {goodsData.usersGoodsPrice}원
-                </GoodsText>
-              </GoodsDonationDiv>
-            </div>
-            <div style={{ width: "100%" }}>
-              <OkorColsebuttonDiv>
-                <div style={{ position: "absolute", top: "85%", right: "10%" }}>
-                  <ApiButton onClick={okButton}>확인</ApiButton>|
-                  <ApiButton
-                    onClick={() => deleteButton(goodsData.usersGoodsId)}
-                  >
-                    취소
-                  </ApiButton>
-                </div>
-              </OkorColsebuttonDiv>
-            </div>
-          </div>
-        ) : (
-          <>
-            {goodsData?.length !== 0 && goodsData ? <></> : <Logo src={logo} />}
-            <Text onChange={getGoodsUrl} value={getGoodsUrlItem} />
-            <div>
-              <GoodsDonationDiv>
-                <p>
-                  상품 이름 : <GoodsNameInput />
-                </p>
-                <p>
-                  후&nbsp; 원 &nbsp; 가 : <GoodsDonationInput />원
-                </p>
-              </GoodsDonationDiv>
-            </div>
-            <div style={{ width: "100%" }}>
-              <OkorColsebuttonDiv>
-                <div>
-                  <ApiButton onClick={registerGoodsButton}>등록하기</ApiButton>
-                </div>
-              </OkorColsebuttonDiv>
-            </div>
-          </>
-        )}
-      </GoodsDiv>
-    </>
-  );
-}
-
-// 수정 상태
-// text url 후원가 상품 이름
-// 수정 삭제하기 버튼 만들어짐
-
-function UpdateGoodsState({
-  isOpen,
-  fetchData,
-  setIsOpen,
-  setFetchData,
-  deleteGoodsRender,
-}) {
-  const [postFilterGoodsData, setPostFilterGoodsData] = useState([]);
-
-  const [editNameText, setEditNameText] = useState("");
-  const [editPriceText, setEditPriceText] = useState("");
-  const [editState, setEditState] = useState({
-    state: false,
-  });
-
-  //상품 가격 수정
-  async function updateGoodsPriceRender(usersGoodsId, usersGoodsName) {
-    const priceData = await updateGoodsPrice(usersGoodsId, usersGoodsName);
-    return priceData;
-  }
-  //상품이름 수정
-  async function updateGoodsNameRender(usersGoodsId, usersGoodsPrice) {
-    const nameData = await updateGoodsname(usersGoodsId, usersGoodsPrice);
-    return nameData;
-  }
-
-  const updateGoodsAllClick = async (usersGoodsId, name, price) => {
-    const priceData = await updateGoodsPriceRender(usersGoodsId, price);
-    if (priceData.success) {
-      setFetchData((prev) =>
-        prev.filter(
-          (goods) => goods.usersGoodsId === priceData.data.usersGoodsId
-        )
-      );
-    }
-    const nameData = await updateGoodsNameRender(usersGoodsId, name);
-    if (nameData.success) {
-      setFetchData((prev) =>
-        prev.filter(
-          (goods) => goods.usersGoodsId === nameData.data.usersGoodsId
-        )
-      );
-    }
-    setIsOpen(false);
-  };
-
-  const goodsDeleteButton = async (id) => {
-    const data = await deleteGoodsRender(id);
-    console.log(data);
-  };
-  const updateGoodsNameChange = (e) => {
-    const value = e.target.value;
-    setEditNameText(value);
-  };
-  const updateGoodsPriceChange = (e) => {
-    const value = e.target.value;
-    setEditPriceText(value);
-  };
-
-  useEffect(() => {
-    const filterGoodsData = fetchData?.filter(
-      (v) => v.usersGoodsId === isOpen.userGoodsId
-    );
-    setEditNameText(filterGoodsData[0].usersGoodsName);
-    setEditPriceText(filterGoodsData[0].usersGoodsPrice);
-    setPostFilterGoodsData(filterGoodsData);
-  }, []);
-
-  return (
-    <>
-      <GoodsDiv>
-        {editState.state
-          ? postFilterGoodsData &&
-            postFilterGoodsData.map((v) => (
-              <div key={v.usersGoodsId}>
-                <GoodsImage url={v.usersGoodsImgUrl} />
-                <div>
-                  <p>
-                    상품 이름 :{" "}
-                    <GoodsNameInput
-                      name="name"
-                      value={editNameText}
-                      onChange={updateGoodsNameChange}
-                    />
-                  </p>
-                  <GoodsDonationDiv>
-                    <p>
-                      후&nbsp; 원 &nbsp; 가 :
-                      <GoodsDonationInput
-                        name="price"
-                        value={editPriceText}
-                        onChange={updateGoodsPriceChange}
-                      />
-                      원
-                    </p>
-                  </GoodsDonationDiv>
-                </div>
-                <div style={{ width: "100%" }}>
-                  <OkorColsebuttonDiv>
-                    <div
-                      style={{ position: "absolute", top: "85%", right: "10%" }}
-                    >
-                      <ApiButton
-                        onClick={() =>
-                          updateGoodsAllClick(
-                            v.usersGoodsId,
-                            editNameText,
-                            editPriceText
-                          )
-                        }
-                      >
-                        수정하기
-                      </ApiButton>
-                      |
-                      <ApiButton
-                        onClick={() => goodsDeleteButton(v.usersGoodsId)}
-                      >
-                        삭제하기
-                      </ApiButton>
-                    </div>
-                  </OkorColsebuttonDiv>
-                </div>
-              </div>
-            ))
-          : postFilterGoodsData &&
-            postFilterGoodsData.map((v) => (
-              <div key={v.usersGoodsId} id={v.usersGoodsId}>
-                <GoodsImage url={v.usersGoodsImgUrl} />
-                <div>
-                  <p>
-                    상품 이름 :{" "}
-                    <GoodsNameInput
-                      value={editNameText}
-                      onChange={updateGoodsNameChange}
-                      name="name"
-                      onFocus={() => setEditState({ state: true })}
-                    />
-                  </p>
-                  <div>
-                    <p
-                      style={{
-                        display: "flex",
-                        alignItems: "center",
-                      }}
-                    >
-                      후&nbsp; 원 &nbsp; 가 :{" "}
-                      <GoodsDonationInput
-                        value={editPriceText}
-                        onFocus={() => setEditState({ state: true })}
-                      />
-                      원
-                    </p>
-                  </div>
-                </div>
-                <div style={{ width: "100%" }}>
-                  <OkorColsebuttonDiv>
-                    <div
-                      style={{ position: "absolute", top: "85%", right: "10%" }}
-                    >
-                      <ApiButton>수정하기</ApiButton>|
-                      <ApiButton
-                        onClick={() => goodsDeleteButton(v.usersGoodsId)}
-                      >
-                        삭제하기
-                      </ApiButton>
-                    </div>
-                  </OkorColsebuttonDiv>
-                </div>
-              </div>
-            ))}
-      </GoodsDiv>
-    </>
-  );
-}
+import UpdateState from "./updateState/UpdateState";
 
 export default function GoodsModal({
   setIsOpen,
@@ -305,7 +42,7 @@ export default function GoodsModal({
             }}
           />
           {isOpen.state === "View" ? (
-            <UpdateGoodsState
+            <UpdateState
               setIsOpen={setIsOpen}
               setFetchData={setFetchData}
               isOpen={isOpen}
@@ -313,7 +50,7 @@ export default function GoodsModal({
               deleteGoodsRender={deleteGoodsRender}
             />
           ) : (
-            <CreateGoodsState
+            <CreateState
               setIsOpen={setIsOpen}
               setGetGoodsUrlItem={setGetGoodsUrlItem}
               postGoodsListRender={postGoodsListRender}
@@ -370,88 +107,4 @@ const TextDiv = styled.div`
   flex-direction: column;
   align-items: center;
   justify-content: space-around;
-`;
-
-const Text = styled.input`
-  width: 479px;
-  height: 29px;
-  background: #eeeeee;
-  border: 1px solid rgba(0, 0, 0, 0.3);
-  border-radius: 20px;
-  margin-top: 10px;
-  padding-left: 10px;
-`;
-
-const GoodsNameInput = styled.input`
-  border: none;
-  background: rgba(228, 230, 232, 0.7);
-  outline: none;
-  width: 400px;
-  height: 20px;
-  margin-bottom: 10px;
-`;
-
-const GoodsDonationInput = styled.input`
-  border: none;
-  background: rgba(228, 230, 232, 0.7);
-  outline: none;
-  width: 100px;
-  height: 20px;
-`;
-
-const GoodsDonationDiv = styled.div`
-  margin-top: 10px;
-`;
-
-const ApiButton = styled.button`
-  border: none;
-  background: none;
-`;
-
-const OkorColsebuttonDiv = styled.div`
-  display: flex;
-  justify-content: center;
-  width: 100%;
-`;
-
-const GoodsDiv = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-  width: 479px;
-`;
-
-const GoodsImage = styled.div`
-  width: 150px;
-  height: 150px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-image: url(${(props) => props.url});
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position: center;
-  margin-bottom: 10px;
-  margin-left: 35%;
-`;
-
-const CreateGoodsImage = styled.div`
-  width: 150px;
-  height: 150px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-image: url(${(props) => props.url});
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position: center;
-  margin-bottom: 10px;
-  margin-left: 30%;
-`;
-
-const GoodsText = styled.p`
-  width: 100%;
-  justify-content: center;
-  display: flex;
 `;
